Fix: refersh data after delegate and refactor
